split index.js into maintainable modules and renamed some functions for clarity
