+++ output/team.html
@@ -1,5 +1,4 @@
-
-    <!DOCTYPE html>
+<!DOCTYPE html>
 <html lang="en">
 
 <head>
@@ -142,4 +141,3 @@
     </div>
 </body>
 </html>
-
